backend/everpro/requirements.txt
amqp==2.6.1
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bottlenose==1.1.8
bs4==0.0.1
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coverage==5.3
Django==3.0.5
django-celery-beat==2.1.0
django-celery-results==1.2.1
django-shortcuts==1.6
django-timezone-field==4.0
djangorestframework==3.12.2
djongo==1.3.3
idna==2.10
importlib-metadata==2.0.0
joblib==0.17.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kombu==4.6.11
lxml==4.6.1
nltk==3.5
nose==1.3.7
prompt-toolkit==3.0.8
psycopg2==2.8.6
pymongo==3.11.0
python-amazon-simple-product-api==2.2.11
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.11
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
tqdm==4.51.0
urllib3==1.26.1
uuid==1.30
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.0
